consolidated all api calls into a class implementing interface
