+++ src/app/chat/[id]/page.tsx
@@ -5,161 +5,56 @@ import { useEffect, useState } from 'react';
 import { useChat } from '@ai-sdk/react';
 import { ChatData } from '@/lib/definitions/types';
 import { chatSchema } from '@/lib/definitions/zod';
+import { clientBotAPI } from '@/lib/api';
 
 export default function Chat() {
+  const api = new clientBotAPI();
+
   const { id: chatId } = useParams();
-  // const [loadingMsgs, setLoadingMsg] = useState<boolean>(true);
-  // const [loadingChat, setLoadingChat] = useState<boolean>(true);
   const [thinking, setThinking] = useState<boolean>(false);
   const [responding, setResponding] = useState<boolean>(false);
   const [chat, setChat] = useState<ChatData | undefined>(undefined)
   
   const [input, setInput] = useState('');
-  const [messages, setMessages] = useState([]);
+  const [messages, setMsgs] = useState([]);
 
   useEffect(() => {
     if (!chatId) return;//maybe redirect here
-    getChat();//these should prob get put into an api class implementing an interface like the tictactoe
-    getMsg();//also maybe this is a good spot for the no waterfall thing?
+
+    const numericChatId = Number(chatId);
+   if (isNaN(numericChatId)) return; 
+
+    api.getChat(numericChatId).then(setChat)
+    api.getMsgs(numericChatId).then(setMsgs);//also maybe this is a good spot for the no waterfall thing?
   }, [chatId]);
 
-  function getChat(){
-    fetch(`/api/chat/${chatId}`)
-    .then((res) => res.json())
-    .then((data) => {
-      if (data) {//this doesnt check for incorrect data
-        const mappedData = {
-          id: data.id,
-          userId: data.user_id,
-          chatName: data.chat_name,
-          createdAt: new Date(data.created_at),
-        };
+  const handleSubmit = async (e) => {
+  e.preventDefault();
+  const newUserMsg = {
+    id: 'temp-user',
+    role: 'user', 
+    content: input,
+  };
 
-        const parsedData = chatSchema.parse(mappedData)
-        setChat(parsedData);
-      } else {
-        console.error('Unexpected data from API:', data);
-        setChat(undefined)
-      }
-    })
-    .catch((err) => {
-      console.error('Failed to fetch chat history:', err);
-      setChat(undefined);
-    });
-  }
-  function getMsg(){
-      fetch(`/api/chat/${chatId}/messages`)
-        .then((res) => res.json())
-        .then((data) => {
-          if (Array.isArray(data)) {
-            setMessages(data);
-          } else {
-            console.error('Unexpected data from API:', data);
-            setMessages([]);
-          }
-        })
-        .catch((err) => {
-          console.error('Failed to fetch chat history:', err);
-          setMessages([]);
-        });
-    }
-
- async function sendMsg(msg: string) {
+  setMsgs(prev => [...prev, newUserMsg]);
   setResponding(true);
-  setThinking(true);
   
   try {
-    const res = await fetch(`/api/chat/${chatId}/messages`, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ msg }),
-    });
-
-    if (!res.ok) {
-      const data = await res.json();
-      console.error('Failed to send message:', data);
-      return;
-    }
-
-    if (!res.body) return;
-
-    setThinking(false);
-    const reader = res.body.pipeThrough(new TextDecoderStream()).getReader();
-    const aiMessageId = `ai-${Date.now()}`;
-    let aiMessageContent = '';
-    let buffer = '';
-
-    try {
-      const newAiMessage = {
-        id: aiMessageId,
-        role: 'assistant',
-        content: '',
-      };
-      setMessages(prev => [...prev, newAiMessage]);
-
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-
-        buffer += value; 
-        const lines = buffer.split('\n');
-        buffer = lines.pop() || '';
-
-        for (const line of lines) {
-          if (line.startsWith('0:')) {
-            try {
-              const content = JSON.parse(line.slice(2));
-              aiMessageContent += content;
-              
-              setMessages(prev => 
-                prev.map(msg =>
-                  msg.id === aiMessageId
-                    ? { ...msg, content: aiMessageContent }
-                    : msg
-                )
-              );
-            } catch (e) {
-              console.error('Failed to parse streaming content:', e);
-            }
-          }
-        }
-      }
-
-      if (aiMessageContent.trim()) {
-        console.log('Saving AI message:', aiMessageContent);
-        
-        const saveResponse = await fetch(`/api/chat/${chatId}/messages/ai`, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({
-            content: aiMessageContent,
-          }),
-        });
-
-        if (saveResponse.ok) {
-          const savedMessage = await saveResponse.json();
-          setMessages(prev =>
-            prev.map(msg =>
-              msg.id === aiMessageId 
-                ? { ...msg, id: savedMessage.id.toString() }
-                : msg
-            )
-          );
-        } else {
-          console.error('Failed to save AI message');
-        }
-      }
-
-    } finally {
-      reader.releaseLock();
-    }
+    await api.postMsg(
+      input, 
+      Number(chatId), 
+      setMsgs, 
+      setThinking,
+      setResponding,
+    );
   } catch (error) {
-    console.error('Error in sendMsg:', error);
+    console.error('Failed to send message:', error);
   } finally {
     setResponding(false);
-    setThinking(false);
+    setInput('');
   }
-}
+};
+
 
   return (
     <div className='chat-container'>
@@ -192,18 +87,7 @@ export default function Chat() {
         </div>
       )}
 
-          <form className ="input-form" onSubmit={e => {
-            e.preventDefault();
-            const newUserMsg = {
-              id: 'temp-user',
-              role: 'user',
-              content: input,
-            };
-
-            setMessages(prev => [...prev, newUserMsg]);
-              sendMsg(input);
-            setInput('');
-          }}>        
+          <form className ="input-form" onSubmit={handleSubmit}>  
           <input
           value={input}
           placeholder="Say something..."
@@ -211,7 +95,8 @@ export default function Chat() {
           disabled={ responding }
           autoFocus
         />
-        <button type="submit" disabled={responding || !input}>
+        <button type="submit" disabled={responding || !chat}>
+          {/* remember to disable on !input */}
           Send
         </button>
       </form>
